Refactored code. Divided slice into /actions and /reducers
